feat(axon_rust-a4w.2): rewrite map_with_sitemap as sitemap-first engine
